requirements.txt
pymilvus
fastapi
towhee==0.6.0
redis
torch
torchvision
uvicorn
opencv-python
matplotlib
pillow
numpy
gdown
python-multipart
keras
numpy
tensorflow
loguru
postgres
